3843 Создал отдельную вьюху на результат по сотрудникам, старую переименовал
