+++ Vodovoz/Views/Complaints/ComplaintResultOfCounterpartyView.cs
@@ -1,0 +1,23 @@
+﻿using QS.Views.GtkUI;
+using Vodovoz.ViewModels.Complaints;
+
+namespace Vodovoz.Views.Complaints
+{
+	[System.ComponentModel.ToolboxItem(true)]
+	public partial class ComplaintResultOfCounterpartyView : TabViewBase<ComplaintResultViewModel>
+	{
+		public ComplaintResultOfCounterpartyView(ComplaintResultViewModel viewModel) : base(viewModel)
+		{
+			this.Build();
+			ConfigureDlg();
+		}
+
+		private void ConfigureDlg()
+		{
+			yentryName.Binding.AddBinding(ViewModel.Entity, e => e.Name, w => w.Text).InitializeFromSource();
+
+			buttonSave.Clicked += (sender, e) => ViewModel.SaveAndClose();
+			buttonCancel.Clicked += (sender, e) => ViewModel.Close(true, QS.Navigation.CloseSource.Cancel);
+		}
+	}
+}
