+++ Vodovoz/Views/Complaints/ComplaintResultOfEmployeesView.cs
@@ -1,0 +1,23 @@
+﻿using QS.Views.GtkUI;
+using Vodovoz.ViewModels.Complaints;
+
+namespace Vodovoz.Views.Complaints
+{
+	[System.ComponentModel.ToolboxItem(true)]
+	public partial class ComplaintResultOfEmployeesView : TabViewBase<ComplaintResultViewModel>
+	{
+		public ComplaintResultOfEmployeesView(ComplaintResultViewModel viewModel) : base(viewModel)
+		{
+			Build();
+			ConfigureDlg();
+		}
+
+		private void ConfigureDlg()
+		{
+			yentryName.Binding.AddBinding(ViewModel.Entity, e => e.Name, w => w.Text).InitializeFromSource();
+
+			buttonSave.Clicked += (sender, e) => ViewModel.SaveAndClose();
+			buttonCancel.Clicked += (sender, e) => ViewModel.Close(true, QS.Navigation.CloseSource.Cancel);
+		}
+	}
+}
